+++ verl/workers/engine/megatron/__init__.py
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .engine_impl import MegatronEngine, MegatronEngineWithLMHead
+from .transformer_impl import MegatronEngine, MegatronEngineWithLMHead
 
 __all__ = ["MegatronEngine", "MegatronEngineWithLMHead"]
